+++ discrete/irregular_interpolate.py
@@ -101,7 +101,7 @@ class IrregularGridInterpolator(object):
 
         # Route all oob points to oob node
         cols[B:] = oob_idx
-        rows[B:] = np.full(num_oob,oob_node)
+        rows[B:] = np.full(num_oob,oob_node,dtype=np.int64)
         data[B:] = np.ones(num_oob)
         
         M = self.num_nodes
